requirements.txt
absl-py==0.9.0
alembic==1.4.1
argcomplete==1.11.1
astor==0.8.1
astunparse==1.6.3
blinker==1.4
boto==2.49.0
boto3==1.12.30
botocore==1.15.30
cachetools==4.1.0
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
coverage==4.0a1
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
durationpy==0.5
Flask==1.0
Flask-Bcrypt==0.6.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.5
Flask-SQLAlchemy==1.0
Flask-Testing==0.4.2
Flask-WTF==0.14.3
flask-talisman==0.7.0
future==0.18.2
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==19.0.0
h5py==2.10.0
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
infinity==1.4
intervals==0.8.1
itsdangerous==0.24
Jinja2==2.11.1
jmespath==0.9.5
kiwisolver==1.1.0
Mako==1.0.0
Markdown==3.2.1
MarkupSafe==0.23
matplotlib==3.2.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.2.0
pip-tools==4.5.1
placebo==0.9.0
protobuf==3.11.3
psycopg2-binary==2.8.4
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.6.1
python-editor==1.0.4
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.43.0
troposphere==2.6.0
urllib3==1.25.8
validators==0.14.2
Werkzeug==0.16.0
wrapt==1.12.1
wsgi-request-logger==0.4.6
WTForms==2.0.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
zappa==0.51.0
zipp==3.1.0
